Correccion de algunos errores :name_badge:
Aun faltan bastantes
El problema principal es que se tuvo que cambiar el tipo de dato de valorDocId de int a bigInt (en sql) y double (en C#)
Cambio y fuera.

+++ Project_1/Controllers/Propiedad_Del_PropietarioController.cs
@@ -14,7 +14,7 @@ namespace Project_1.Controllers
     {   
         [Route("PropiedadDelPropietario/Despropiar/{numeroFinca}/{valorDocId}")]
         [HttpPost]
-        public bool Despropiar(int numeroFinca, int valorDocId)
+        public bool Despropiar(int numeroFinca, double valorDocId)
         {
             var propiedadDelPropietario = new Propiedad_Del_Propietario()
             {
@@ -29,7 +29,7 @@ namespace Project_1.Controllers
 
         [Route("PropiedadDelPropietario/Apropiar/{numeroFinca}/{valorDocId}")]
         [HttpPost]
-        public string Apropiar(int numeroFinca, int valorDocId)
+        public string Apropiar(int numeroFinca, double valorDocId)
         {
             if (!ModelState.IsValid)
             {
